Fix clobbering of breaklist and replstring
After a bit of digging after the root cause for #73, I finally fond
this.

This was introduced in "Imported from conserver-8.1.19.tar.gz" (b94c8967)
and was probably just a typo.

When found #73 I saw that the same kind of "bug" happened to breaklist
to so I fixed that one to.

Signed-off-by: Anton Lundin <glance@ac2.se>